# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my application",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/properties": {
            "get": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Get a list of properties",
                "operationId": "getAllProperties",
                "responses": {
                    "200": {
                        "description": "A list of properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Property"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Create a property",
                "operationId": "storeProperty",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manufacturing"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "House"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Florida"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "455000"
                                    },
                                    "manufacturing": {
                                        "type": "integer",
                                        "example": "2018"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/property/{id}": {
            "get": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Get a single property",
                "operationId": "getPropertyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the property",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Update a property",
                "operationId": "updateProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the property",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "address",
                                    "price",
                                    "manufacture"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "House"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Florida"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "456000"
                                    },
                                    "manufacture": {
                                        "type": "integer",
                                        "example": "2020"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Property not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Delete a property",
                "operationId": "deleteProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Property": {
                "title": "Property",
                "required": [
                    "id",
                    "type",
                    "address",
                    "price",
                    "manufacturing"
                ],
                "properties": {
                    "id": {
                        "description": "Property ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "description": "Property's type",
                        "type": "string"
                    },
                    "address": {
                        "description": "Property's address",
                        "type": "string"
                    },
                    "price": {
                        "description": "Property's price",
                        "type": "integer"
                    },
                    "manufacturing": {
                        "description": "Property's manufacturing",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}